Fix the odd placing of ح dots in حر
I was convinced that this was intentional and as they knew better than
me I shall not change it. It turned out that they did not have
a pre-composed glyph with dots on the ح for that ligature, so the odd
placing is a result of the mechanical limitation of composing dots
"manually". Now feeling liberated, I adjusted the dots to a more
sensible position. Too much talking for such a small fix :)