Get ride of glyph insertion
OpenType does not support glyph insertion but I was working around this
by using contextual + multiple substitution, e.g. to insert C between
AB I do a contextual substitution for AB that replaces A->AC, but this
seems to make problems with mark positioning in a way incompatible
between Uniscribe and Pango; Pango will position A marks on the second
component; C, so I redid it by replacing B->CB, but Uniscibe positions
the dots on first component, oh well!

So now I just create a new AB glyph that contains references of A and B.