Workaround a Pango bug
The first subtable in positioning lookup used in contextual positions
can contain one positioning record at most or no positioning will be
applied at all, but no such issues with other subtables, so I add all
other records in a second subtable for now. Need to check if this is
really a bug and report it.