An attempt to get around OOo bugs
OpenOffice has a serious bug with horizontal glyph positioning in RTL
direction; it seems to calculate the positions correctly and the word
length is as expected (i.e. equals the total of glyph advance widths and
offsets) but when the glyphs are actually drawn the offsets are ignored
and the glyphs extend out of the word bounds. This have been manifested
by broken kerning in the past[1], but it is more serious with Amiri
since we relay heavily in anchor marks for connecting glyphs and the
result was a total mess.

To get around that, as a temporary measure, I've changed the side bearing
of all anchored glyphs so that the x offset is always zero, this
improves the situation considerably, but the issue with kerning remains
unsolved.

[1] http://www.openoffice.org/issues/show_bug.cgi?id=89286