Make sure tashkil above dots is always applied
If one types:

  <base><mark above><mark below>

it will be resolved to:

  <base><dot above><dot below><mark above><mark below>

and so neither <mark above> nor <mark below> will be applied to its
respective dot because of the other mark or dot getting in the way. To
fix this we define two mark sets for marks & dots above and marks & dots
below, and in the taskil above dot lookups all marks are ignored except
those in the respective set.