Dissociate alternate different alternate dots
The alternate dot on ر and ف is used in cases where a certain character
requires a dot in an unusual position like the centered dot below ف
compared to standard one below its head. I was using one dot for that
purpose and for contextual dot positioning, but I'm going to implement
the later in a more clean way that does not require an alternate glyph.