Fix Tashkil above س/ص in سح/صح
For the new fixed vertical position of the Tashkil, base glyphs need not
to move vertically, which can happen sometimes because of cursive
anchors (if the glyph moved vertically, the tashkil anchor will move
with it and thus no longer "fixed"). Since we don't actually need the
vertical move for anything, I just moved the glyphs so that the cursive
anchor will not move them.