Initial step in dropping "dynamic dots"
The concept though very elegant theoretically have been hindering
further development of the font in practice; OpenType is not ready for
this type of development and I've been hitting so many limitations and
implementation bugs that I gave up.

Instead the "dynamic dots" idea will be only used during development and
glyphs will be composed once for all.

This commit adds the (pre-)composed glyphs, feature files will be
adjusted in the following commits.