Reimplement kerning as contextual positioning
OpenType kerning can either be done as simple kerning pairs where every
two kerned glyphs are listed individually, or kerning classes where
glyphs are grouped into classes and classes are kerned as a whole,
however in kerning classes accepts only a single value that controls the
advance width of the first glyph in the pair which rules out its use in
RTl kerning (which requires changing the advance with of two glyphs in
the pair, because engines re-order the glyph then lay them from left to
right). As a result, I was using classes in the feature file but with
`enum` keyword which instructs FontForge to convert them into simple
pairs, so I can avoid the need of listing every individual pair.

This worked fine with the old font layout, but after recent development
the number of glyphs grow by 7 folds and the number of kerning pairs
became insanely high; around half a million kerning pairs!

Now I've reimplemented kerning as contextual positioning which is much
saner and avoids the need to change the advance width of both glyphs
(one only needs to change the advance width of the *second* glyph).