Support Quranic alternate dagger alef
In Quranic orthography (at least in Egypt and most of Arabic world)
there are two forms of dagger alef; a superscript one that is a
combining mark and a spacing one that is regular alef letter, the later
is not encoded in Unicode yet. The spacing dagger alef is almost always
preceded by a fatha mark and we use this fact to replace the combing
dagger alef by a spacing one when preceded by fatha, a hack but should
do the job until the proper character get encoded.