Fill most of Presentation Forms-A and B blocks
Built using tools/build_compat.py script, but it is easier to do it once
and keep the result in the repository than keep doing it with each
build, also hb-shape is not always available (e.g. building the fonts
down stream). The script should be rerun whenever there is change that
would affect its output.