Use mark classes instead of sets for overline
fontTools does not understand mark sets and when we transcode the web
version with it (as it produces smaller file size), the GSUB table of
the resulting font is rejected by OTS, breaking the font with Firefox
and probably chrome. Since we only have one set in the font, there is no
practical difference between it and mark classes.

Fixes last part of:
https://sourceforge.net/p/amiri/feature-requests/14/