Make existing Gaf-like glyphs an alternate form
My Gaf is inspired by an old Turkish manuscript where it feels more like
a Naskh letter than the common form used in today fonts that looks more
like a retrofitted Nastaliq letter, but because the horizontal stroke of
my Gaf is hanging on the right, it can easily colloid with preceding
tall glyph and the final form colloids with vowels on the preceding
letter (not an issue for Persian where vowels are not used, but it is
an issue for Arabic though Gaf is not that widely used).

So I’ll be designing a new form, and now I’m moving the existing one to
an alternate glyph to make room for the new default.