Change the inverted damma
In Indonesia, the inverted damma is really just a 180° rotated damma,
but in Urdu it has a morphed, more natural shape. I was using the Urdu
shape, but other Naskh fonts use the rotated damma and the Urdu shape is
unintelligible in Indonesia. So now I moved the Urdu shape to a locale
specific variant and make a rotated damma the default.

Fixes http://sourceforge.net/p/amiri/feature-requests/11/