Fix coloring medial small alef
The tatweel part of the glyph should not be colored, so we now split the
glyph into two layers and color them separately.

Ideally, the tatweel part should just take the text color and the spec
supports doing so, but Mozilla implementation rejects the font in this
case, so for the time being I’m just setting it to black.